Move Add Discovered Reader to dedicated service
