+++ dds/src/implementation/domain_participant_backend/services/mod.rs
@@ -1,5 +1,6 @@
 pub mod data_reader_service;
 pub mod data_writer_service;
+pub mod discovery_service;
 pub mod domain_participant_service;
 pub mod event_service;
 pub mod message_service;
